Camera position attribute created
'bounds' is now a fixed value that represents the region on the screen where the camera will be rendered. Position is the offset from its limit's origin.
It looks like it's working but there are some bugs that need to be fixed before going to the next step, which is keeping limit always updated.